Faster heap iteration
Most of the time spent parsing a heap dump was in
Metadata.instantiateWrapperFor, which calls
BasicTypeDataBase.findDynamicTypeForAddress. That code does a very slow
(roughly linear) search through all C++ classes to try to find one with the
same vtable pointer.

Instead, use metadata available in the (C++) Oop/Klass in order to determine
the correct (Java) subclasses of Oop and Klass for each heap object. Also, keep
around a reference to the original Klass, so that future calls to getKlass
return quickly.