Support for TLABs
This seems to fix the last of my issues with misses during probing. It turns
out that the TLAB areas are actually inside the eden space, which results in
long strings of zeros in the middle of the heap. Filtering these out reduces
the miss rate to practically zero.

I also switched to using RangeSet from Guava instead of a home-baked
AddressRange.