Performance improvements, mostly
Attempt to reduce the number of objects allocated while probing for heap
objects. The idea is to reduce the probability of garbage collection, and
reduce the amount that the dumping thread interferes with the heap contents.

Byte buffers are allocated up front instead of on-demand. I also believe there
were some subtle bugs in the buffer cache that I weeded out here. If a page
became unmapped after the SelfInferior was initialized, it was possible that it
would fail to read, but the stale contents of the page would continue to be
used in the AddressSpace. Also enlarge the page cache based on some sample
usage statistics that I gathered.

Greatly speed up the statistical dump when the heap contains large objects that
have a reasonably high change of being randomly hit more than once.
Unfortunately, the code to detect these objects requires allocations, and it
somewhat counteracts the first point.